feat: followUser, unfollowUser, updateUser, getProfile
